// src/components/cdk/cdk-empty-state-one/models/cdk-empty-state-one.model.ts
export interface ICDKEmptyStateLiterals {
	/** Título del empty state */
	mainTitle: string;

	/** Descripción */
	text: string;

	/** Botón*/
	btn: string;

	/** Titulo card*/
	smallTitle: string;

	/** Descripcion de la card*/
	desc: string;

	/** Imagen izquierda*/
	imgLeftSrc: string;

	/** Imagen derecha*/
	imgRightSrc: string;

	/** Imagen card*/
	imgSrc: string;
}